+++ common/Rhyolite/Route.hs
@@ -30,8 +30,6 @@ import Data.Text.Encoding
 import Network.HTTP.Types.URI (parseQuery, renderQuery)
 import Network.URI
 
-import Rhyolite.Request.Common (decodeValue')
-
 class Monad m => MonadRoute r m | m -> r where
   routeToUrl :: r -> m URI
 
@@ -115,12 +113,12 @@ addSubdomain sub uri = uri { uriAuthority = fmap (\auth -> auth { uriRegName = s
 getDefaultParam :: FromJSON b => Map BS.ByteString (Maybe BS.ByteString) -> Maybe b
 getDefaultParam params = do
   Just v <- Map.lookup (encodeUtf8 "x") params
-  decodeValue' (LBS.fromStrict v)
+  decodeStrict' v
 
 decodeRoute :: (FromJSON r) => T.Text -> Maybe r
 decodeRoute t = do
   Just v <- Map.lookup (encodeUtf8 "x") (Map.fromList (parseQuery (encodeUtf8 t)))
-  decodeValue' (LBS.fromStrict v)
+  decodeStrict' v
 
 uriToRouteEnv
   :: URI